@@ -26,6 +26,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import java.io.Serial;
 import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
@@ -38,6 +39,7 @@ import java.util.UUID;
 @Entity
 public final class User extends BasicUserDetails {
 
+    @Serial
     private static final long serialVersionUID = 1L;
 
     /**
@@ -152,10 +154,9 @@ public final class User extends BasicUserDetails {
         if (that == null) {
             return false;
         }
-        if (!(that instanceof User)) {
+        if (!(that instanceof final User other)) {
             return false;
         }
-        final var other = (User) that;
         return id.equals(other.id);
     }
 
